AUT-783: Move TestClientHelper to shared
So it can be consumed by the DynamoClientService
